Mass rewrite of file structure -- had to modify makefile to accommodate, hope nothing is actually broken
