+++ src/HarmonicBayes/HarmonicFunctions.cpp
