+++ src/HarmonicBayes/harmonicFit.cpp
