--- src/harmonicFit.h
@@ -1,9 +1,0 @@
-#pragma once
-#include "data.h"
-#include "JSL.h"
-#include "NegativeBinomial.h"
-#include "plotting.h"
-#include "settings.h"
-#include "Transitions.h"
-#include "HarmonicFunctions.h"
-void HarmonicFit(Data & d, const Settings & settings);
